moved combinatorial_Data.lifter to lifter
